moving generic multiprocessing routine to utils. general code cleanup
