+++ hexrd/WPPF.py
@@ -5,6 +5,7 @@ from hexrd.imageutil import snip1d
 from hexrd.crystallography import PlaneData
 from hexrd.material import Material
 from hexrd.valunits import valWUnit
+from hexrd.utils.multiprocess_generic import GenericMultiprocessing
 from hexrd import spacegroup as SG
 from hexrd import symmetry, symbols, constants
 import hexrd.resources
@@ -20,7 +21,6 @@ import time
 import h5py
 from pathlib import Path
 from pylab import plot, ginput, show, axis, close, title
-from multiprocessing import Queue, Process, cpu_count
 
 class Parameters:
     ''' ========================================================================================================
@@ -2455,9 +2455,6 @@ def extract_intensities(polar_view,
     ========================================================================================================
     ========================================================================================================
     '''
-    # get the maximum number of processes that will be started
-    nprocs = cpu_count()
-    print("# of cpu = ", nprocs, "running on all of them.")
 
     # prepare the data file to distribute suing multiprocessing
     data_inp_list = []
@@ -2479,105 +2476,9 @@ def extract_intensities(polar_view,
     'bkgmethod': bkgmethod
     }
 
-    P = UptownFunc()
+    P = GenericMultiprocessing()
     results = P.parallelise_function(data_inp_list, single_azimuthal_extraction, **kwargs)
     return results
-
-        # # loop over different inputs and pass to single_azimuthal_extraction
-        # processes = []
-        # q_in = Queue(1)
-        # q_out = Queue()
-
-        # for i in range(nprocs):
-        #     pass_args = [func, q_in, q_out]
-        #     # pass_args.extend(args)
-
-        #     p = Process(target=self._func_queue,\
-        #                 args=tuple(pass_args),\
-        #                 kwargs=kwargs)
-
-        #     processes.append(p)
-
-        # for p in processes:
-        #     p.daemon = True
-        #     p.start()
-
-        # sent = [q_in.put((i, data_inp_list[i])) for i in range(len(data_inp_list))]
-        # [q_in.put((None, None)) for _ in range(nprocs)]
-
-        # results = [[] for i in range(n)]
-        # for i in range(len(sent)):
-        #     index, res = q_out.get()
-        #     results[index] = res
-
-        # # wait until each processor has finished
-        # [p.join() for p in processes]
-
-        # # reorder results
-        # return results
-
-        # for i in range(polar_view.size[0]):
-        #     args = (data_inp_list[i],)
-
-        #     P = Process(target=single_azimuthal_extraction,
-        #                 args=args,
-        #                 kwargs=kwargs)
-        #     result = 
-
-class UptownFunc:
-    def __init__(self):
-        pass
-
-    def _func_queue(self, func, q_in, q_out, *args, **kwargs):
-        """ Retrive processes from the queue """
-        while True:
-            pos, var = q_in.get()
-            if pos is None:
-                break
-
-            res = func(var, *args, **kwargs)
-            q_out.put((pos, res))
-        return
-
-    def parallelise_function(self, var, func, *args, **kwargs):
-        """ Split evaluations of func across processors """
-        n = len(var)
-
-        processes = []
-        q_in = Queue(1)
-        q_out = Queue()
-
-        nprocs = cpu_count()
-
-        for i in range(nprocs):
-            pass_args = [func, q_in, q_out]
-            # pass_args.extend(args)
-
-            p = Process(target=self._func_queue,\
-                        args=tuple(pass_args),\
-                        kwargs=kwargs)
-
-            processes.append(p)
-
-        for p in processes:
-            p.daemon = True
-            p.start()
-
-        # put items in the queue
-        sent = [q_in.put((i, var[i])) for i in range(n)]
-        [q_in.put((None, None)) for _ in range(nprocs)]
-
-        # get the results
-        results = [[] for i in range(n)]
-        for i in range(len(sent)):
-            index, res = q_out.get()
-            results[index] = res
-
-        # wait until each processor has finished
-        [p.join() for p in processes]
-
-        # reorder results
-        return results
 
 def single_azimuthal_extraction(expt_spectrum,
                         params=None,
